CI: Define release/2411 branch policy (#318)
This change updates the mirroring policy for the new release/2411 branch
to automatically mirror commits to the internal ADO repository. It also
disables building and publishing of the openvmm guide since we want that
to always be coming from the main branch. In a future flowey enhancement
we should investigate making these types of jobs have support for
conditional execution.

---------

Co-authored-by: Ben Hillis <benhill@ntdev.microsoft.com>