Add support for xsyncing rust-toolchain.toml files (#354) (#361)
We use rust-toolchain.toml files in our release branches to ensure that
all development on them is locked to a specific version of Rust. Add
support to xsync for ensuring the overlay matches oss. This is being
taken to 2411 right now, and then a rust-toolchain.toml file will be
added for 2411, and we should be all set.