Cherry-pick: vmbus_server: call unstick_channels to mitigate the lost synic issue (#405)
Cherry-pick restore path fix for lost synic issue.

This PR addresses an issue where the synic event was lost during the
save process. To mitigate this issue during the restore process, we have
introduced a new field lost_synic_bug_fixed in the SavedState structure.
The lost_synic_bug_fixed field will indicate whether the issue was fixed
in save code path. During the restore process, we check if the field is
not set then unstick_channels() function is called to mitigate the
issue.

In below kmsg traces, yellow line shows we check the field to see if
need call unstick_channels and green line shows we did hit the issue
where the rings were not empty and need signal the guest:



![image](https://github.com/user-attachments/assets/222a6131-2085-4ec8-a067-c68ee91e67be)

---------

Co-authored-by: Juan Tian <juantian@microsoft.com>