openhcl: log how large the saved state buffer is for servicing (#400) (#431)
We know how large it is, but log it so it's discoverable. 

In the future, we are probably moving to a model where we do not send
the saved state to the host. This will help us decide how large to size
a reserved memory range inside OpenHCL.

Cherry pick of #400