openhcl_boot: set private pool size to 1 page for ak cert (#508)
This is a release/2411 only change.

Set the private pool to 1 page on non isolated VMs to support AK cert
renewal. This is safe to do, because we do not support live DMA on these
ranges across servicing, removing the need save restoring these
allocations, as we do not allow this request to be in flight while
servicing.