whp: Refactor get/set property to improve safety and fix a UB (#801)
Fixes an issue in the WHP code that caused the WHV_ARM64_IC_PARAMETERS
struct to get trampled on when using a release build. This was caused by
the use of an unsafe block to return a reference to a struct that went
out of scope. In debug builds, the compiler didn't reuse this memory
right away, but in the release build it did. Now the original struct
outlives the function that uses it.

---------

Co-authored-by: John Starks <jostarks@microsoft.com>