vmbus_async: don't panic on empty external data (#621) (#802)
Don't `panic` when a GpaDirect packet contains a pointer to empty GPA
ranges. This data is guest controlled and not a programmer error, so
`panic` is not correct. Consumers of `read_external_ranges` already
handle errors gracefully, so this is not a significant change for them.

As a drive-by code improvement, also change some cases of `#[from]` to
`#[source]` in error definitions, and add the requisite `.map_err`
changes.

Co-authored-by: Matt Kurjanowicz <mattkur@microsoft.com>