Record a livedump of OpenHCL when hitting a guest or uefi watchdog timeout (#1093) (#1099)
The aim here is to produce as much information as we can to in debugging
these timeouts after-the-fact in prod. This dump should end up in watson
whenever these timeouts are hit, which gives us something to investigate
besides just "we timed out". This does not yet implement our idea of
including inspect output in these dumps, that will be a follow up.