virt_mshv_vtl: tdx: don't double inject events (#1139) (#1140)
The TDX backend tracks the next pending interruption as a kind of shadow
register on the VMCS's interruption information field. It needs to clear
this pending interruption on the next exit after the L2 runs to avoid
reinjecting the interruption a second time.

When kernel exit offloading is enabled, some exits are handled
internally, without returning to user mode. In this case, the TDX
backend does not observe the exit and so does not clear the pending
interruption shadow register. If the next return from kernel mode is not
due to an exit but due to an EINTR return or TDX_PENDING_INTERRUPT, then
it will assume the L2 did not run and will reinject the pending
interruption a second time.

To avoid this, disable the kernel exit offloading when there is a
pending interruption.

A future change could probably simplify this by avoiding the use of a
shadow register and just using the VMCS as the source of truth.

Backport of #1139

Co-authored-by: John Starks <jostarks@microsoft.com>