Preserve TPM AK cert across boots for resiliency and Improve attestation error log and telemetry (#1198)
This PR ports #1120 and #1124 to 2411 release branch. Changes include
the following

- Preserves the AK cert (if TPM seeds are not refreshed) across boots
that increases the resiliency in case that the boot-time AK cert request
succeeds in the first boot but fails in the subsequent boot.
- Ensure the AK cert index is always re-created with consistent
properties (index size, platform-created, write permission per-boot auth
value)
- Enable VMM tests to cover this code change
- Improve the error log when WrappedKey response is required but invalid
and the telemetry on recovering VMGS KEK with hardware-based sealing
key.
- Improve existing attestation logs

---------

Signed-off-by: Ming-Wei Shih <mishih@microsoft.com>