Tx GDMA error handling on 2411 (#1231)
The net_mana::tx_poll() does not properly handle a CQE of type
CQE_TX_GDMA_ERR. GDMA_ERR is sent from the Hardware to indicate that the
queue has been disabled. Once the queue is disabled, no further traffic
can be sent out from the Guest. The proper way to handle the Gdma error
is to recreate/restart the tx queue.

Changes:

* Adding logging to help track and triage these issues. A Gdma error can
be returned due to any number of problems, so the only way to determine
the cause is by logging the vendor error and properties of the packet.
* Modifying net_mana::tx_poll to return an error that indicates the
queue should be restarted.
* Modifying netvsp to handle the TxError::TryReturn
* If process_endpoint_tx sees a TryReturn error, it completes any
pending tx packets before the queues are restarted.
* If main_loop returns a TryReturn error, Coordinator is sent a Restart
message, which will restart all the queues.
* 
Notes:

* I'm leveraging the existing code to restart all queues, which is
disruptive, but better than being unable to send traffic.
* The underlying cause of the Gdma error(s) still need to be addressed.
Openvmm netvsp is crafting packets in such a way to cause hardware
errors, the logging should help in determining the problem.
* While we observed crashes in openvmm after a Gdma Error when the
packet was None, I could not recreate a repro in the lab. It remains
unclear if these changes will address the crash.