netvsp: Mismatched Indirection Table error on servicing ( #771) 2411 (#1327)
OpenVMM Netvsp driver has a hardcoded assumption that the RSS
Indirection Table size will not change during servicing. There are rare
cases where capabilities may be modified during re-advertisement. This
occurred when MANA switched from giving a flat 128 to scaling based on
the number of vcpu cores during a FHR. The result was a failed servicing
operation.

Dynamic changes to the RSS Table Size are difficult because there is no
mechanism for informing the Guest to re-query RSS capabilities, which it
assumes do not change during its lifetime. Reducing the Indirection
Table size can cause unexpected and difficult to investigate issues with
network performance and processor overloading. In Azure, VMs are
allocated the minimum number of indirection table entries for VTL2 VF at
startup. And we have a commitment that the number will not be further
reduced during servicing. However, we do not want OpenVMM to fail
servicing operations if the Nic advertises fewer RSS table entries. The
code has been modified to log such events, but then ignore the change in
size and proceed.

Increasing the Indirection Table size during servicing may be useful.
The code has been modified to extend the RSS Indirection Table and
replay the existing entries to fill up the new spaces.

Tested and working in a lab environment.

Added unit tests.

Note: Testing Requires Ubuntu 22.04 / kernel 6.8-azure. Some older
versions of Ubuntu do not have a netvsc that honors the RSS table size.