tpm: mitigate corruption due to corrupted TPM NVRAM blob (#1527)
Mitigate TPM corruption due to previous VMs having a 16K TPM NVRAM
reported as 32K, and commited bad state to the vTPM NVRAM.

This involves the following:

For every 16K TPM NVRAM, walk the dynamic section and truncate the last
header if it points to data past the end.

Additionally, run the following mitigation steps for 16K NVRAM:
1. Check for a 4K bytes AK cert nv index. 1. This VM needs to be
mitigated. 2. Undefine the 4K AK cert to save space. 3. Attempt to write
a 1 byte mitigation platform marker, which can fail. 4. Attempt to write
a just-sized platform ak cert.
2. Else, check for a mitigated marker or no platform cert 1. Log that
this vm is mitigated, and if an ak cert is present or not
3. Else, check for an owner cert 1. Log that this VM is in the expected
state

Co-authored-by: Chris Oo <cho@microsoft.com>