openhcl/CVM: disable mnf support (#1712) (#1715)
MNF support does not work on CVM. Disable it for now to unbreak
scenarios such as MANA nic relay, while we root cause the actual issue.
Tracked by #1711.

Onboard a test that forces MNF for TDX and other flavors. Note that this
test is not complete, as we do need to verify within the guest that MNF
is actually present. But at a minimum, we can at least see that toggling
this property does not break the guest. Tracked by #1713.

Cherry pick of #1712

Co-authored-by: Steven Malis <137308034+smalis-msft@users.noreply.github.com>