fuzz_nvme_driver: Use payload memory for transfer buffer (#1770)
Using the first half of the guest memory, as was done previously,
overlaps with the memory space allocated to the emulated NVMe device.
This overlap means that the IO targets can overwrite memory precious to
the device, such as the IO submission or completion queues.

Of important note: the fuzzer is only targeting a portion of the stack.
The ranges are trusted at this point in the `nvme_driver`.