vmcore/vm_topology: support pmu interrupt on aarch64 (#1776) (#1797)
On AArch64, the Performance Monitor Unit (PMU) is supposed to be
supported by every platform. Add this information to the vm's topology,
and correctly report a configured value in the MADT via the GICC
structure. Onboard a test to verify that Linux sees the correct
interrupts.

Hyper-V and WHP support a hardcoded value of 0x17, so for now hardcode
that value on those platforms. A follow up change will correctly report
this value via a `pmu` device tree node, but take this more minimal
change to backport to the release/2505 branch.

Although macOS also supports this interrupt with the same value of 0x17,
enabling that did not cause Linux to work as expected, so more
investigation there is needed.

This fixes xperf on Windows and perf on Linux which rely on this being
present.

#1775 

Backport of #1776