nvme_manager: multithreaded (#1854)
Cherry pick to release/2505

### underhill_core: factory for nvme devices + tests for nvme_manager
(#1787)

Add unit tests for the existing code in `nvme_manager.rs`. This requires
a minor refactoring: push the code to create NVMe drivers into a factory
that the tests can then mock.

These basic tests already highlight the performance problems seen in
production: the GetNamespace path is not concurrent.

This is one part of the broader work effort.

### underhill_core: nvme_manager: make it multithreaded #1763

This PR addresses serialization in the existing underhill_core:
nvme_manager. This serialization proves to be the bottleneck when
performing a runtime servicing operation with multiple NVMe devices.

This change leverages mesh to create a two-level hierarchy:
- The existing `NvmeManager` API surface is the top level. The idea is
that this keeps track of the NVMe devices that are in some state of
being created, and
- A new `NvmeDriverManager` that manages the lifecycle of a single NVMe
device.

Most NVMe devices have one namespace, but our cloud scale scenario
requires supporting multiple namespaces per NVMe device. It's okay to
serialize multiple calls to the same device, since the most expensive
portion is loading the driver.